requirements.txt
# Perp Lobster - Dependencies

# Core exchange SDK
hyperliquid-python-sdk>=0.21.0

# Ethereum/crypto
eth-account>=0.11.0

# Web framework (for dashboard)
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.35.0

# Date parsing
python-dateutil>=2.8.0

# HTTP
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# AI analysis features (optional)
anthropic>=0.39.0
